[P] Fix acceleration data types

+++ MapView/domain/accelerometer.py
@@ -2,6 +2,6 @@ from dataclasses import dataclass
 
 @dataclass
 class Accelerometer:
-  x: int
-  y: int
-  z: int
+  x: float
+  y: float
+  z: float
